exception handlers are now passed to rx.App instead of rx.Config
