+++ tests/test_state.py
@@ -1421,11 +1421,13 @@ def test_error_on_state_method_shadow():
 
 
 @pytest.mark.asyncio
-async def test_state_with_invalid_yield(capsys):
+async def test_state_with_invalid_yield(capsys, mock_app):
     """Test that an error is thrown when a state yields an invalid value.
 
     Args:
         capsys: Pytest fixture for capture standard streams.
+        mock_app: Mock app fixture.
+
     """
 
     class StateWithInvalidYield(BaseState):
@@ -1669,7 +1671,7 @@ def mock_app(monkeypatch, state_manager: StateManager) -> rx.App:
     Returns:
         The app, after mocking out prerequisites.get_app()
     """
-    app = App(state=TestState)
+    app = App(state=TestState, app_name="mock_app")
 
     app_module = Mock()
